refactor: split inference by semantic ownership
