+++ jazz-next/src/JazzNext/Compiler/TypeInference/Capabilities.hs
@@ -1,36 +1,175 @@
 {-# LANGUAGE OverloadedStrings #-}
 
 module JazzNext.Compiler.TypeInference.Capabilities
-  ( applyCapabilityFacts,
+  ( SignaturePayloadType (..),
+    applyCapabilityFacts,
+    addInferredEqualityClassConstraintIfVisible,
+    addUnpreservedInferredMethodConstraintErrors,
+    applyTypeSchemePrimitiveConstraints,
+    builtinDollarOperatorExpr,
     capabilityFactsFromState,
+    checkImplMethodBodies,
+    classMethodPayloadToExpressionType,
+    defaultBindingLiteralTypes,
+    defaultLiteralTypes,
+    deferExplicitConstraints,
+    deferExplicitConstraintsWithFacts,
     enterModuleCapabilityScope,
+    expressionTypeToRuntimeHint,
+    expressionTypeToRuntimeHintWithVariables,
+    finalizeDeferredExplicitConstraintsAt,
     flushCurrentModuleCapabilityFacts,
+    freeTypeVariables,
+    freeTypeVariablesInEnv,
+    freeTypeVariablesInTypeSchemeConstraints,
+    freshTypeVars,
     generatedEqualityCapabilityName,
     importModuleCapabilityFacts,
+    inferQualifiedMethodApplication,
+    instantiateQualifiedMethodType,
+    instantiateTypeSchemeConstraint,
+    instantiateTypeSchemePrimitiveConstraint,
     mergeCapabilityFacts,
+    qualifiedMethodClassIsVisible,
+    replaceTypeVariables,
+    resolveTypeSchemeConstraint,
     restoreCapabilityFacts,
     seedFacts,
     seedStatementCapabilityFact,
     typeSchemeDefiningFactsFromState,
     typeSchemeReferencedCapabilityFacts,
+    constraintSignatureTypeToExpressionType,
+    dedupeTypeSchemeConstraints,
+    signaturePayloadToSignatureType,
+    signatureTypeToExpressionType,
+    structuralRuntimeEqualityType,
     updateRootModuleBaselineFacts
   ) where
 
-import Data.List (foldl' )
+import Control.Applicative ((<|>))
+import Data.List (foldl')
+import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
+import Data.Maybe (isNothing)
+import Data.Set (Set)
 import qualified Data.Set as Set
 import Data.Text (Text)
+import qualified Data.Text as Text
 import JazzNext.Compiler.AST
-  ( ClassMethodSignature (..), ConstraintSignatureType, ImplMethod (..), Statement (..) )
+  ( CaseArm (..),
+    ClassMethodSignature (..),
+    ConstraintSignatureType (..),
+    Expr (..),
+    ImplMethod (..),
+    NumericType (..),
+    SignatureConstraint (..),
+    SignaturePayload (..),
+    SignatureType (..),
+    Statement (..)
+  )
+import JazzNext.Compiler.BuiltinCatalog
+  ( BuiltinResolutionMode,
+    numericTypeFromName,
+    numericTypeIsIntegral,
+    renderNumericTypeName
+  )
 import JazzNext.Compiler.CapabilityFacts
-  ( concreteConstraintArgument, concreteImplFactClassName, concreteImplFactKey, qualifiedMethodKey, splitQualifiedMethodKey )
-import JazzNext.Compiler.Identifier (identifierText)
+  ( concreteConstraintArgument,
+    concreteImplFactClassName,
+    concreteImplFactKey,
+    constraintImplFactKey,
+    constraintFunctionArgumentTypes,
+    constraintSignatureTypeVariableNamesInOrder,
+    constraintSignatureAliasVariants,
+    constraintSignatureTypeContainsClassParameter,
+    constraintSignatureTypesCompatible,
+    identifierLooksLikeTypeVariable,
+    normalizeConstraintSignatureName,
+    qualifiedMethodKey,
+    renderConstraintSignatureType,
+    signaturePayloadConstraintType,
+    splitQualifiedMethodKey,
+    substituteClassMethodSignature
+  )
+import JazzNext.Compiler.Diagnostics
+  ( SourceSpan,
+    setDiagnosticPrimarySpan
+  )
+import JazzNext.Compiler.Identifier
+  ( identifierText,
+    mkIdentifier,
+    qualifiedIdentifierText
+  )
 import JazzNext.Compiler.Name
-  ( GeneratedNameKind (..), Name (..), NameNamespace (..), renderName )
+  ( GeneratedNameKind (..),
+    Name (..),
+    NameNamespace (..),
+    qualifiedMemberName,
+    renderName,
+    sourceName
+  )
+import JazzNext.Compiler.TypeInference.Diagnostics
+  ( InferExprFn,
+    addTypeError,
+    annotateNewErrorsWithPrimarySpan,
+    duplicateConstraintName,
+    mkAmbiguousDeferredConstraintError,
+    mkAmbiguousQualifiedMethodBodyError,
+    mkAmbiguousQualifiedMethodBodyForArgumentsError,
+    mkApplyTypeError,
+    mkExplicitConstraintArityError,
+    mkImplMethodMissingClassMethodError,
+    mkImplMethodTypeMismatchError,
+    mkInvalidQualifiedMethodSignatureError,
+    mkMissingClassMethodError,
+    mkMissingExplicitConstraintClassError,
+    mkMissingExplicitConstraintImplFactError,
+    mkMissingImplMethodBodyError,
+    mkNoMatchingQualifiedMethodBodyError,
+    mkTypeSchemeNumericConstraintError,
+    mkTypeSchemeStrictEqualityConstraintError
+  )
 import JazzNext.Compiler.TypeInference.State
-  ( DeclarationState (..), InferState (..), ModuleInferenceState (..), inferClassFacts, inferClassMethodSignatures, inferConcreteImplFacts, inferConcreteImplMethods, inferCurrentModuleLocalCapabilityFacts, inferCurrentModulePath, inferGeneratedEqualityClassFacts, inferModuleCapabilityFacts )
+  ( DeclarationState (..),
+    DeferredExplicitConstraint (..),
+    InferState (..),
+    InferenceOutput (..),
+    ModuleInferenceState (..),
+    inferClassFacts,
+    inferClassMethodSignatures,
+    inferConcreteImplFacts,
+    inferConcreteImplMethods,
+    inferCurrentModuleLocalCapabilityFacts,
+    inferCurrentModulePath,
+    inferDataTypes,
+    inferDeferredExplicitConstraints,
+    inferErrorCount,
+    inferGeneratedEqualityClassFacts,
+    inferInferredClassConstraints,
+    inferModuleCapabilityFacts,
+  )
+import JazzNext.Compiler.TypeInference.Solver
+  ( addStrictEqualityTypeVarConstraint,
+    constrainNumericOperatorType,
+    freshTypeVar,
+    integerLiteralRangeFitsNumericType,
+    resolveType,
+    supportsRuntimeEqualityType,
+    unifyTypes
+  )
 import JazzNext.Compiler.TypeInference.Types
-  ( ClassMethodType (..), ImplMethodType (..), ScopeCapabilityFacts (..), TypeSchemeConstraint (..), emptyScopeCapabilityFacts )
+  ( ClassMethodType (..),
+    ConstructorArgumentType (..),
+    ExpressionType (..),
+    ImplMethodType (..),
+    ScopeCapabilityFacts (..),
+    TypeBinding (..),
+    TypeEnv,
+    TypeScheme (..),
+    TypeSchemeConstraint (..),
+    TypeSchemePrimitiveConstraint (..),
+    emptyScopeCapabilityFacts
+  )
 
 capabilityFactsFromState :: InferState -> ScopeCapabilityFacts
 capabilityFactsFromState state =
@@ -335,3 +474,1635 @@ modifyDeclarationState update state =
 modifyModuleInferenceState :: (ModuleInferenceState -> ModuleInferenceState) -> InferState -> InferState
 modifyModuleInferenceState update state =
   state {inferModule = update (inferModule state)}
+
+modifyInferenceOutput :: (InferenceOutput -> InferenceOutput) -> InferState -> InferState
+modifyInferenceOutput update state =
+  state {inferOutput = update (inferOutput state)}
+
+builtinDollarOperatorExpr :: TypeEnv -> Expr -> Bool
+builtinDollarOperatorExpr env expr =
+  case expr of
+    EOperatorValue "$" -> True
+    EVar name ->
+      case Map.lookup name env of
+        Just (BuiltinOperatorAliasTypeBinding "$") -> True
+        Just (OperatorAliasSchemeTypeBinding "$" _) -> True
+        _ -> False
+    _ -> False
+
+qualifiedMethodClassIsVisible :: Text -> InferState -> Bool
+qualifiedMethodClassIsVisible methodKey state =
+  case splitQualifiedMethodKey methodKey of
+    Just (capabilityName, _) -> Map.member capabilityName (inferClassFacts state)
+    Nothing -> False
+
+inferQualifiedMethodApplication ::
+  InferExprFn ->
+  BuiltinResolutionMode ->
+  TypeEnv ->
+  InferState ->
+  Text ->
+  [Expr] ->
+  (Maybe ExpressionType, InferState)
+inferQualifiedMethodApplication inferExpression builtinMode env state methodKey argumentExprs =
+  let (argumentTypes, stateAfterArguments) =
+        inferQualifiedMethodArguments inferExpression builtinMode env state argumentExprs
+   in case sequence argumentTypes of
+        Nothing -> (Nothing, stateAfterArguments)
+        Just typedArgumentTypes ->
+          resolveQualifiedMethodApplicationType
+            methodKey
+            env
+            stateAfterArguments
+            (zip argumentExprs typedArgumentTypes)
+
+inferQualifiedMethodArguments ::
+  InferExprFn ->
+  BuiltinResolutionMode ->
+  TypeEnv ->
+  InferState ->
+  [Expr] ->
+  ([Maybe ExpressionType], InferState)
+inferQualifiedMethodArguments inferExpression builtinMode env state argumentExprs =
+  let (reversedTypes, finalState) = foldl' step ([], state) argumentExprs
+   in (reverse reversedTypes, finalState)
+  where
+    step (typesAcc, stateAcc) argumentExpr =
+      let (argumentType, stateAfterArgument) =
+            inferExpression builtinMode env stateAcc argumentExpr
+       in (argumentType : typesAcc, stateAfterArgument)
+
+checkImplMethodBodies ::
+  InferExprFn ->
+  BuiltinResolutionMode ->
+  TypeEnv ->
+  InferState ->
+  Name ->
+  [ConstraintSignatureType] ->
+  [ImplMethod] ->
+  InferState
+checkImplMethodBodies inferExpression builtinMode env state capabilityName arguments methods =
+  case arguments of
+    [implTarget]
+      | concreteConstraintArgument implTarget,
+        not (implMethodNamesHaveDuplicates methods) ->
+          let implMethodEnv stateForBindings =
+                Map.union env (currentImplMethodBindings implTarget stateForBindings)
+              checkMethod stateAcc (ImplMethod methodName methodSpan methodExpr) =
+                let methodKey = qualifiedMethodKey capabilityName methodName
+                 in case Map.lookup methodKey (inferClassMethodSignatures stateAcc) of
+                      Nothing ->
+                        addTypeError
+                          stateAcc
+                          (mkImplMethodMissingClassMethodError methodKey methodSpan)
+                      Just classMethodType ->
+                        let (maybeExpectedType, stateAfterExpectedType) =
+                              qualifiedMethodSignatureType
+                                methodKey
+                                classMethodType
+                                (ImplMethodType implTarget)
+                                stateAcc
+                         in case maybeExpectedType of
+                              Nothing ->
+                                stateAfterExpectedType
+                              Just expectedType ->
+                                let (maybeMethodType, rawStateAfterMethod) =
+                                      inferExprTypeWithExpected inferExpression
+                                        builtinMode
+                                        (implMethodEnv stateAcc)
+                                        stateAfterExpectedType
+                                        expectedType
+                                        methodExpr
+                                    stateAfterMethod =
+                                      annotateNewErrorsWithPrimarySpan methodSpan stateAfterExpectedType rawStateAfterMethod
+                                    stateAfterMethodCheck =
+                                      case maybeMethodType of
+                                        Just methodType ->
+                                          case unifyTypes expectedType methodType stateAfterMethod of
+                                            Just unifiedState -> unifiedState
+                                            Nothing ->
+                                              addTypeError
+                                                stateAfterMethod
+                                                ( mkImplMethodTypeMismatchError
+                                                    methodKey
+                                                    methodSpan
+                                                    (resolveType stateAfterMethod expectedType)
+                                                    (resolveType stateAfterMethod methodType)
+                                                )
+                                        Nothing ->
+                                          stateAfterMethod
+                                 in finalizeDeferredExplicitConstraintsAt
+                                      methodSpan
+                                      stateAfterExpectedType
+                                      stateAfterMethodCheck
+           in foldl' checkMethod state methods
+    _ -> state
+  where
+    implMethodNamesHaveDuplicates :: [ImplMethod] -> Bool
+    implMethodNamesHaveDuplicates implMethods =
+      let methodNames = map (\(ImplMethod methodName _ _) -> identifierText methodName) implMethods
+       in length methodNames /= Set.size (Set.fromList methodNames)
+
+    currentImplMethodBindings :: ConstraintSignatureType -> InferState -> TypeEnv
+    currentImplMethodBindings implTarget stateForBindings =
+      Map.fromList
+        [ (qualifiedMemberName capabilityName methodName, PlainTypeBinding methodType)
+          | ImplMethod methodName _ _ <- methods,
+            let methodKey = qualifiedMethodKey capabilityName methodName,
+            Just (ClassMethodType classParameter methodSignature) <- [Map.lookup methodKey (inferClassMethodSignatures stateForBindings)],
+            Just methodType <- [classMethodPayloadToExpressionType stateForBindings classParameter implTarget methodSignature]
+        ]
+
+inferExprTypeWithExpected ::
+  InferExprFn ->
+  BuiltinResolutionMode ->
+  TypeEnv ->
+  InferState ->
+  ExpressionType ->
+  Expr ->
+  (Maybe ExpressionType, InferState)
+inferExprTypeWithExpected inferExpression builtinMode env state expectedType expr =
+  case (resolveType state expectedType, expr) of
+    (TFunctionType argumentType resultType, ELambda parameterName bodyExpr) ->
+      let extendedEnv =
+            Map.insert parameterName (PlainTypeBinding argumentType) env
+          (bodyType, stateAfterBody) =
+            inferExprTypeWithExpected inferExpression builtinMode extendedEnv state resultType bodyExpr
+       in case bodyType of
+            Just inferredBodyType ->
+              ( Just
+                  ( TFunctionType
+                      (resolveType stateAfterBody argumentType)
+                      inferredBodyType
+                  ),
+                stateAfterBody
+              )
+            Nothing -> (Nothing, stateAfterBody)
+    _ -> inferExpression builtinMode env state expr
+
+addUnpreservedInferredMethodConstraintErrors ::
+  SourceSpan ->
+  TypeEnv ->
+  InferState ->
+  InferState ->
+  ExpressionType ->
+  Set Int ->
+  InferState
+addUnpreservedInferredMethodConstraintErrors spanValue env statementStartState state statementResultType schemeVariables =
+  foldl'
+    addUnpreservedClassConstraintError
+    ( foldl'
+        addUnpreservedMethodError
+        (foldl' addUnpreservedConcreteMethodConstraintError state droppedConcreteMethodConstraints)
+        droppedAmbiguousMethodKeys
+    )
+    droppedClassConstraints
+  where
+    droppedClassConstraints =
+      dedupeTypeSchemeConstraints
+        [ TypeSchemeInferredConstraint constraintName argumentType
+          | TypeSchemeInferredConstraint constraintName argumentType <-
+              newInferredClassConstraints statementStartState state,
+            not (inferredConstraintTargetPreserved state schemeVariables argumentType),
+            not (inferredConstraintTargetStillVisibleInEnv state env argumentType),
+            inferredConstraintTargetConcrete state argumentType
+              || ( not statementIntroducedErrors
+                     && inferredConstraintTargetEscapesResult state statementResultType argumentType
+                 )
+        ]
+
+    droppedMethodConstraints =
+      dedupeTypeSchemeConstraints
+        [ TypeSchemeMethodConstraint constraintName methodKey argumentType
+          | TypeSchemeMethodConstraint constraintName methodKey argumentType <-
+              newInferredClassConstraints statementStartState state,
+            not (inferredConstraintTargetPreserved state schemeVariables argumentType),
+            not (inferredConstraintTargetStillVisibleInEnv state env argumentType),
+            not (concreteInferredMethodConstraintSatisfied state constraintName methodKey argumentType)
+        ]
+
+    droppedConcreteMethodConstraints =
+      [ methodConstraint
+        | methodConstraint@(TypeSchemeMethodConstraint _ _ argumentType) <- droppedMethodConstraints,
+          inferredConstraintTargetConcrete state argumentType
+      ]
+
+    droppedAmbiguousMethodKeys =
+      Set.toList
+        ( Set.fromList
+            [ methodKey
+              | TypeSchemeMethodConstraint _ methodKey argumentType <- droppedMethodConstraints,
+                not (inferredConstraintTargetConcrete state argumentType)
+            ]
+        )
+
+    addUnpreservedMethodError stateAcc methodKey =
+      addTypeError
+        stateAcc
+        (setDiagnosticPrimarySpan spanValue (mkAmbiguousQualifiedMethodBodyError methodKey))
+
+    addUnpreservedConcreteMethodConstraintError stateAcc constraint =
+      annotateNewErrorsWithPrimarySpan
+        spanValue
+        stateAcc
+        ( resolveDeferredExplicitConstraint
+            stateAcc
+            (typeSchemeConstraintToDeferredExplicitConstraint (capabilityFactsFromState state) (capabilityFactsFromState state) constraint)
+        )
+
+    addUnpreservedClassConstraintError stateAcc constraint =
+      annotateNewErrorsWithPrimarySpan
+        spanValue
+        stateAcc
+        ( resolveDeferredExplicitConstraint
+            stateAcc
+            (typeSchemeConstraintToDeferredExplicitConstraint (capabilityFactsFromState state) (capabilityFactsFromState state) constraint)
+        )
+
+    statementIntroducedErrors =
+      inferErrorCount state > inferErrorCount statementStartState
+
+newInferredClassConstraints :: InferState -> InferState -> [TypeSchemeConstraint]
+newInferredClassConstraints previousState state =
+  take newConstraintCount (inferInferredClassConstraints state)
+  where
+    previousConstraintCount = length (inferInferredClassConstraints previousState)
+    currentConstraintCount = length (inferInferredClassConstraints state)
+    newConstraintCount = max 0 (currentConstraintCount - previousConstraintCount)
+
+inferredConstraintTargetPreserved :: InferState -> Set Int -> ExpressionType -> Bool
+inferredConstraintTargetPreserved state schemeVariables argumentType =
+  let targetType = resolveType state argumentType
+      targetVariables = freeTypeVariables targetType
+   in not (Set.null targetVariables)
+        && targetVariables `Set.isSubsetOf` schemeVariables
+
+inferredConstraintTargetConcrete :: InferState -> ExpressionType -> Bool
+inferredConstraintTargetConcrete state argumentType =
+  let resolvedArgumentType = defaultLiteralTypes (resolveType state argumentType)
+   in Set.null (freeTypeVariables resolvedArgumentType)
+        && case expressionTypeToRuntimeHint resolvedArgumentType of
+          Just _ -> True
+          Nothing -> False
+
+inferredConstraintTargetStillVisibleInEnv :: InferState -> TypeEnv -> ExpressionType -> Bool
+inferredConstraintTargetStillVisibleInEnv state env argumentType =
+  let targetType = resolveType state argumentType
+      targetVariables = freeTypeVariables targetType
+      environmentVariables = freeTypeVariablesInEnv state env
+   in not (Set.null targetVariables)
+        && targetVariables `Set.isSubsetOf` environmentVariables
+
+inferredConstraintTargetEscapesResult :: InferState -> ExpressionType -> ExpressionType -> Bool
+inferredConstraintTargetEscapesResult state statementResultType argumentType =
+  let targetVariables = freeTypeVariables (resolveType state argumentType)
+      resultVariables = freeTypeVariables (resolveType state statementResultType)
+   in not (Set.null targetVariables)
+        && not (Set.null (Set.intersection targetVariables resultVariables))
+
+concreteInferredMethodConstraintSatisfied :: InferState -> Text -> Text -> ExpressionType -> Bool
+concreteInferredMethodConstraintSatisfied state constraintName methodKey argumentType =
+  let resolvedArgumentType = resolveType state argumentType
+      facts = capabilityFactsFromState state
+   in Set.null (freeTypeVariables resolvedArgumentType)
+        && concreteInferredMethodConstraintHasUniqueCandidate facts state constraintName methodKey resolvedArgumentType
+
+concreteInferredMethodConstraintHasUniqueCandidate :: ScopeCapabilityFacts -> InferState -> Text -> Text -> ExpressionType -> Bool
+concreteInferredMethodConstraintHasUniqueCandidate facts state constraintName methodKey argumentType =
+  case satisfyingMethodHints of
+    [] -> False
+    [_] -> True
+    _
+      | expressionTypeContainsUncommittedIntegerLiteral argumentType ->
+          uniqueExactRuntimeCandidateHint state argumentType satisfyingMethodHints
+      | otherwise -> True
+  where
+    satisfyingMethodHints =
+      [ argumentHint
+        | argumentHint <- inferredConstraintCandidateRuntimeHints facts state (Just methodKey) argumentType,
+          concreteImplFactExists constraintName argumentHint facts,
+          concreteImplMethodBodyExists methodKey argumentHint facts
+      ]
+
+uniqueExactRuntimeCandidateHint :: InferState -> ExpressionType -> [ConstraintSignatureType] -> Bool
+uniqueExactRuntimeCandidateHint state argumentType candidateHints =
+  case
+      [ candidateHint
+        | candidateHint <- candidateHints,
+          constraintSignatureTypeExactlyMatchesExpressionType state candidateHint argumentType
+      ] of
+    [candidateHint] ->
+      not (constraintSignatureTypeContainsList candidateHint)
+    _ -> False
+
+dedupeTypeSchemeConstraints :: [TypeSchemeConstraint] -> [TypeSchemeConstraint]
+dedupeTypeSchemeConstraints =
+  foldr insertIfMissing []
+  where
+    insertIfMissing constraint constraints
+      | constraint `elem` constraints = constraints
+      | otherwise = constraint : constraints
+
+resolveTypeSchemeConstraint :: InferState -> TypeSchemeConstraint -> TypeSchemeConstraint
+resolveTypeSchemeConstraint state constraint =
+  case constraint of
+    TypeSchemeConstraint constraintName argumentType ->
+      TypeSchemeConstraint constraintName (resolveType state argumentType)
+    TypeSchemeInferredConstraint constraintName argumentType ->
+      TypeSchemeInferredConstraint constraintName (resolveType state argumentType)
+    TypeSchemeMethodConstraint constraintName methodKey argumentType ->
+      TypeSchemeMethodConstraint constraintName methodKey (resolveType state argumentType)
+
+freeTypeVariablesInTypeSchemeConstraints :: [TypeSchemeConstraint] -> Set Int
+freeTypeVariablesInTypeSchemeConstraints constraints =
+  Set.unions (map freeTypeVariablesInTypeSchemeConstraint constraints)
+
+freeTypeVariablesInTypeSchemeConstraint :: TypeSchemeConstraint -> Set Int
+freeTypeVariablesInTypeSchemeConstraint constraint =
+  case constraint of
+    TypeSchemeConstraint _ argumentType ->
+      freeTypeVariables argumentType
+    TypeSchemeInferredConstraint _ argumentType ->
+      freeTypeVariables argumentType
+    TypeSchemeMethodConstraint _ _ argumentType ->
+      freeTypeVariables argumentType
+
+resolveTypeSchemePrimitiveConstraint :: InferState -> TypeSchemePrimitiveConstraint -> TypeSchemePrimitiveConstraint
+resolveTypeSchemePrimitiveConstraint state primitiveConstraint =
+  case primitiveConstraint of
+    TypeSchemeNumericConstraint numericConstraint argumentType ->
+      TypeSchemeNumericConstraint numericConstraint (resolveType state argumentType)
+    TypeSchemeStrictEqualityConstraint argumentType ->
+      TypeSchemeStrictEqualityConstraint (resolveType state argumentType)
+
+freeTypeVariablesInTypeSchemePrimitiveConstraints :: [TypeSchemePrimitiveConstraint] -> Set Int
+freeTypeVariablesInTypeSchemePrimitiveConstraints primitiveConstraints =
+  Set.unions (map freeTypeVariablesInTypeSchemePrimitiveConstraint primitiveConstraints)
+
+freeTypeVariablesInTypeSchemePrimitiveConstraint :: TypeSchemePrimitiveConstraint -> Set Int
+freeTypeVariablesInTypeSchemePrimitiveConstraint primitiveConstraint =
+  case primitiveConstraint of
+    TypeSchemeNumericConstraint _ argumentType -> freeTypeVariables argumentType
+    TypeSchemeStrictEqualityConstraint argumentType -> freeTypeVariables argumentType
+
+freeTypeVariablesInEnv :: InferState -> TypeEnv -> Set Int
+freeTypeVariablesInEnv state =
+  Set.unions . map (freeTypeVariablesInBinding state) . Map.elems
+
+freeTypeVariablesInBinding :: InferState -> TypeBinding -> Set Int
+freeTypeVariablesInBinding state binding =
+  case binding of
+    PlainTypeBinding expressionType ->
+      freeTypeVariables (resolveType state expressionType)
+    SchemeTypeBinding typeScheme ->
+      Set.difference
+        ( Set.unions
+            [ freeTypeVariables (resolveType state (schemeResultType typeScheme)),
+              freeTypeVariablesInTypeSchemeConstraints (map (resolveTypeSchemeConstraint state) (schemeClassConstraints typeScheme)),
+              freeTypeVariablesInTypeSchemePrimitiveConstraints (map (resolveTypeSchemePrimitiveConstraint state) (schemePrimitiveConstraints typeScheme))
+            ]
+        )
+        (schemeQuantifiedVariables typeScheme)
+    OperatorAliasSchemeTypeBinding _ typeScheme ->
+      Set.difference
+        ( Set.unions
+            [ freeTypeVariables (resolveType state (schemeResultType typeScheme)),
+              freeTypeVariablesInTypeSchemeConstraints (map (resolveTypeSchemeConstraint state) (schemeClassConstraints typeScheme)),
+              freeTypeVariablesInTypeSchemePrimitiveConstraints (map (resolveTypeSchemePrimitiveConstraint state) (schemePrimitiveConstraints typeScheme))
+            ]
+        )
+        (schemeQuantifiedVariables typeScheme)
+    BuiltinAliasTypeBinding {} -> Set.empty
+    BuiltinOperatorAliasTypeBinding {} -> Set.empty
+    ConstructorTypeBinding _ _ argumentTypes ->
+      Set.unions (map (freeTypeVariablesInConstructorArgument state) argumentTypes)
+
+freeTypeVariablesInConstructorArgument :: InferState -> ConstructorArgumentType -> Set Int
+freeTypeVariablesInConstructorArgument state argumentType =
+  case argumentType of
+    ConstructorArgumentMonomorphic expressionType ->
+      freeTypeVariables (resolveType state expressionType)
+    ConstructorArgumentParameter {} -> Set.empty
+    ConstructorArgumentFresh -> Set.empty
+
+freeTypeVariables :: ExpressionType -> Set Int
+freeTypeVariables expressionType =
+  case expressionType of
+    TIntType -> Set.empty
+    TIntegerLiteralType {} -> Set.empty
+    TFloatType -> Set.empty
+    TNumericType {} -> Set.empty
+    TBoolType -> Set.empty
+    TListType elementType ->
+      freeTypeVariables elementType
+    TTupleType elementTypes ->
+      Set.unions (map freeTypeVariables elementTypes)
+    TDataType _ typeArguments ->
+      Set.unions (map freeTypeVariables typeArguments)
+    TFunctionType inputType outputType ->
+      Set.union (freeTypeVariables inputType) (freeTypeVariables outputType)
+    TVarType typeVar ->
+      Set.singleton typeVar
+
+replaceTypeVariables :: Map Int ExpressionType -> ExpressionType -> ExpressionType
+replaceTypeVariables replacements expressionType =
+  case expressionType of
+    TIntType -> TIntType
+    TIntegerLiteralType literalRange -> TIntegerLiteralType literalRange
+    TFloatType -> TFloatType
+    TNumericType numericType -> TNumericType numericType
+    TBoolType -> TBoolType
+    TListType elementType ->
+      TListType (replaceTypeVariables replacements elementType)
+    TTupleType elementTypes ->
+      TTupleType (map (replaceTypeVariables replacements) elementTypes)
+    TDataType typeName typeArguments ->
+      TDataType typeName (map (replaceTypeVariables replacements) typeArguments)
+    TFunctionType inputType outputType ->
+      TFunctionType
+        (replaceTypeVariables replacements inputType)
+        (replaceTypeVariables replacements outputType)
+    TVarType typeVar ->
+      Map.findWithDefault expressionType typeVar replacements
+
+-- | Pending type signature state mirrors analyzer adjacency rules while
+-- carrying the normalized declaration type for the next binding.
+instantiateTypeSchemeConstraint :: Map Int ExpressionType -> TypeSchemeConstraint -> TypeSchemeConstraint
+instantiateTypeSchemeConstraint replacements constraint =
+  case constraint of
+    TypeSchemeConstraint constraintName argumentType ->
+      TypeSchemeConstraint constraintName (replaceTypeVariables replacements argumentType)
+    TypeSchemeInferredConstraint constraintName argumentType ->
+      TypeSchemeInferredConstraint constraintName (replaceTypeVariables replacements argumentType)
+    TypeSchemeMethodConstraint constraintName methodKey argumentType ->
+      TypeSchemeMethodConstraint constraintName methodKey (replaceTypeVariables replacements argumentType)
+
+instantiateTypeSchemePrimitiveConstraint :: Map Int ExpressionType -> TypeSchemePrimitiveConstraint -> TypeSchemePrimitiveConstraint
+instantiateTypeSchemePrimitiveConstraint replacements primitiveConstraint =
+  case primitiveConstraint of
+    TypeSchemeNumericConstraint numericConstraint argumentType ->
+      TypeSchemeNumericConstraint numericConstraint (replaceTypeVariables replacements argumentType)
+    TypeSchemeStrictEqualityConstraint argumentType ->
+      TypeSchemeStrictEqualityConstraint (replaceTypeVariables replacements argumentType)
+
+applyTypeSchemePrimitiveConstraints :: [TypeSchemePrimitiveConstraint] -> InferState -> InferState
+applyTypeSchemePrimitiveConstraints primitiveConstraints state =
+  foldl' applyPrimitiveConstraint state primitiveConstraints
+  where
+    applyPrimitiveConstraint stateAcc primitiveConstraint =
+      case primitiveConstraint of
+        TypeSchemeNumericConstraint numericConstraint argumentType ->
+          case constrainNumericOperatorType numericConstraint argumentType stateAcc of
+            Just nextState -> nextState
+            Nothing ->
+              addTypeError
+                stateAcc
+                (mkTypeSchemeNumericConstraintError numericConstraint (resolveType stateAcc argumentType))
+        TypeSchemeStrictEqualityConstraint argumentType ->
+          case resolveType stateAcc argumentType of
+            TVarType typeVar ->
+              addStrictEqualityTypeVarConstraint typeVar stateAcc
+            resolvedType
+              | supportsRuntimeEqualityType stateAcc resolvedType ->
+                  stateAcc
+              | otherwise ->
+                  addTypeError stateAcc (mkTypeSchemeStrictEqualityConstraintError resolvedType)
+
+deferExplicitConstraints :: [TypeSchemeConstraint] -> InferState -> InferState
+deferExplicitConstraints explicitConstraints state =
+  deferExplicitConstraintsWithFacts (capabilityFactsFromState state) (capabilityFactsFromState state) explicitConstraints state
+
+deferExplicitConstraintsWithFacts :: ScopeCapabilityFacts -> ScopeCapabilityFacts -> [TypeSchemeConstraint] -> InferState -> InferState
+deferExplicitConstraintsWithFacts facts structuralFacts explicitConstraints state
+  | null explicitConstraints = state
+  | otherwise =
+      modifyInferenceOutput
+        ( \output ->
+            output
+              { outputDeferredConstraints =
+                  inferDeferredExplicitConstraints state
+                    ++ map (typeSchemeConstraintToDeferredExplicitConstraint facts structuralFacts) explicitConstraints
+              }
+        )
+        state
+
+typeSchemeConstraintToDeferredExplicitConstraint :: ScopeCapabilityFacts -> ScopeCapabilityFacts -> TypeSchemeConstraint -> DeferredExplicitConstraint
+typeSchemeConstraintToDeferredExplicitConstraint facts structuralFacts constraint =
+  case constraint of
+    TypeSchemeConstraint constraintName argumentType ->
+      DeferredExplicitConstraint
+        { deferredConstraintName = constraintName,
+          deferredMethodKey = Nothing,
+          deferredWasInferred = False,
+          deferredArgumentType = argumentType,
+          deferredVisibleFacts = facts,
+          deferredStructuralFacts = structuralFacts
+        }
+    TypeSchemeInferredConstraint constraintName argumentType ->
+      DeferredExplicitConstraint
+        { deferredConstraintName = constraintName,
+          deferredMethodKey = Nothing,
+          deferredWasInferred = True,
+          deferredArgumentType = argumentType,
+          deferredVisibleFacts = facts,
+          deferredStructuralFacts = structuralFacts
+        }
+    TypeSchemeMethodConstraint constraintName methodKey argumentType ->
+      DeferredExplicitConstraint
+        { deferredConstraintName = constraintName,
+          deferredMethodKey = Just methodKey,
+          deferredWasInferred = True,
+          deferredArgumentType = argumentType,
+          deferredVisibleFacts = facts,
+          deferredStructuralFacts = structuralFacts
+        }
+
+finalizeDeferredExplicitConstraintsAt :: SourceSpan -> InferState -> InferState -> InferState
+finalizeDeferredExplicitConstraintsAt spanValue statementStartState state =
+  annotateNewErrorsWithPrimarySpan
+    spanValue
+    state
+    (resolveStatementDeferredExplicitConstraints statementStartState state)
+
+resolveStatementDeferredExplicitConstraints :: InferState -> InferState -> InferState
+resolveStatementDeferredExplicitConstraints statementStartState state =
+  foldl' resolveDeferredExplicitConstraint stateWithoutStatementConstraints statementConstraints
+  where
+    priorConstraints = inferDeferredExplicitConstraints statementStartState
+    currentConstraints = inferDeferredExplicitConstraints state
+    statementConstraints =
+      drop (length priorConstraints) currentConstraints
+    stateWithoutStatementConstraints =
+      modifyInferenceOutput
+        (\output -> output {outputDeferredConstraints = priorConstraints})
+        state
+
+resolveDeferredExplicitConstraint :: InferState -> DeferredExplicitConstraint -> InferState
+resolveDeferredExplicitConstraint state deferredConstraint =
+  let unresolvedArgumentType =
+        resolveType state argumentType
+      resolvedArgumentType =
+        defaultLiteralTypes unresolvedArgumentType
+   in
+    if not (Set.null (freeTypeVariables unresolvedArgumentType))
+      then addTypeError state (mkAmbiguousDeferredConstraintError inferredConstraint constraintName resolvedArgumentType)
+      else
+        case Map.lookup constraintName (scopeClassFacts facts) of
+          Nothing ->
+            addTypeError state (mkMissingExplicitConstraintClassError constraintName)
+          Just classArity
+            | classArity /= 1 ->
+                addTypeError state (mkExplicitConstraintArityError constraintName classArity)
+            | otherwise ->
+                case constraintRuntimeHintsForDeferred facts state inferredConstraint constraintName maybeMethodKey unresolvedArgumentType of
+                  [] ->
+                    addTypeError state (mkAmbiguousDeferredConstraintError inferredConstraint constraintName resolvedArgumentType)
+                  argumentHints ->
+                    let implFactHints =
+                          filter
+                            (constraintImplFactExistsForDeferred facts inferredConstraint constraintName)
+                            argumentHints
+                        methodBodyHints methodKey =
+                          filter
+                            (\argumentHint -> concreteImplMethodBodyExists methodKey argumentHint facts)
+                            implFactHints
+                        ambiguousMethodBodyHints methodKey =
+                          inferredConstraint
+                            && expressionTypeContainsUncommittedIntegerLiteral unresolvedArgumentType
+                            && length (methodBodyHints methodKey) > 1
+                            && not (uniqueExactRuntimeCandidateHint state unresolvedArgumentType (methodBodyHints methodKey))
+                        renderedImplFactKey =
+                          constraintName <> "(" <> renderConstraintSignatureType (head argumentHints) <> ")"
+                     in case maybeMethodKey of
+                          Nothing
+                            | not (null implFactHints) ->
+                                state
+                            | inferredConstraint
+                                && inferredEqualityConstraintCanUseStructuralRuntimeEquality state structuralFacts maybeMethodKey constraintName resolvedArgumentType ->
+                                state
+                            | otherwise ->
+                                addTypeError state (mkMissingExplicitConstraintImplFactError renderedImplFactKey)
+                          Just methodKey
+                            | null implFactHints ->
+                                addTypeError state (mkMissingExplicitConstraintImplFactError renderedImplFactKey)
+                            | ambiguousMethodBodyHints methodKey ->
+                                addTypeError state (mkAmbiguousQualifiedMethodBodyError methodKey)
+                            | not (null (methodBodyHints methodKey)) ->
+                                state
+                            | otherwise ->
+                                addTypeError state (mkMissingImplMethodBodyError methodKey)
+  where
+    constraintName = deferredConstraintName deferredConstraint
+    maybeMethodKey = deferredMethodKey deferredConstraint
+    inferredConstraint = deferredWasInferred deferredConstraint
+    argumentType = deferredArgumentType deferredConstraint
+    facts = deferredVisibleFacts deferredConstraint
+    structuralFacts = deferredStructuralFacts deferredConstraint
+
+expressionTypeContainsUncommittedIntegerLiteral :: ExpressionType -> Bool
+expressionTypeContainsUncommittedIntegerLiteral expressionType =
+  case expressionType of
+    TIntegerLiteralType {} -> True
+    TListType elementType ->
+      expressionTypeContainsUncommittedIntegerLiteral elementType
+    TTupleType elementTypes ->
+      any expressionTypeContainsUncommittedIntegerLiteral elementTypes
+    TDataType _ typeArguments ->
+      any expressionTypeContainsUncommittedIntegerLiteral typeArguments
+    TFunctionType argumentType resultType ->
+      expressionTypeContainsUncommittedIntegerLiteral argumentType
+        || expressionTypeContainsUncommittedIntegerLiteral resultType
+    _ -> False
+
+constraintRuntimeHintsForDeferred ::
+  ScopeCapabilityFacts ->
+  InferState ->
+  Bool ->
+  Text ->
+  Maybe Text ->
+  ExpressionType ->
+  [ConstraintSignatureType]
+constraintRuntimeHintsForDeferred facts state inferredConstraint _ maybeMethodKey argumentType
+  | inferredConstraint =
+      inferredConstraintCandidateRuntimeHints facts state maybeMethodKey argumentType
+  | otherwise =
+      case expressionTypeToRuntimeHint (defaultLiteralTypes argumentType) of
+        Just argumentHint -> [argumentHint]
+        Nothing -> []
+
+constraintImplFactExistsForDeferred :: ScopeCapabilityFacts -> Bool -> Text -> ConstraintSignatureType -> Bool
+constraintImplFactExistsForDeferred facts inferredConstraint constraintName argumentHint =
+  if inferredConstraint
+    then concreteImplFactExists constraintName argumentHint facts
+    else Set.member implFactKey (scopeConcreteImplFacts facts)
+  where
+    implFactKey = constraintName <> "(" <> renderConstraintSignatureType argumentHint <> ")"
+
+inferredConstraintCandidateRuntimeHints :: ScopeCapabilityFacts -> InferState -> Maybe Text -> ExpressionType -> [ConstraintSignatureType]
+inferredConstraintCandidateRuntimeHints facts state maybeMethodKey argumentType =
+  dedupeConstraintSignatureTypes (defaultHint ++ methodCandidateHints)
+  where
+    defaultHint =
+      case expressionTypeToRuntimeHint (defaultLiteralTypes argumentType) of
+        Just argumentHint -> [argumentHint]
+        Nothing -> []
+
+    methodCandidateHints =
+      case maybeMethodKey of
+        Nothing -> []
+        Just methodKey ->
+          [ implTarget
+            | ImplMethodType implTarget <- Map.findWithDefault [] methodKey (scopeConcreteImplMethods facts),
+              constraintSignatureTypeMatchesExpressionType state implTarget argumentType
+          ]
+
+dedupeConstraintSignatureTypes :: [ConstraintSignatureType] -> [ConstraintSignatureType]
+dedupeConstraintSignatureTypes =
+  go Set.empty
+  where
+    go _ [] = []
+    go seen (signatureType : rest)
+      | Set.member rendered seen = go seen rest
+      | otherwise = signatureType : go (Set.insert rendered seen) rest
+      where
+        rendered = renderConstraintSignatureType signatureType
+
+constraintSignatureTypeMatchesExpressionType :: InferState -> ConstraintSignatureType -> ExpressionType -> Bool
+constraintSignatureTypeMatchesExpressionType state signatureType expressionType =
+  case (signatureType, resolveType state expressionType) of
+    (ConstraintTypeName signatureName, TIntegerLiteralType literalRange) ->
+      case numericTypeFromConstraintSignatureName (identifierText signatureName) of
+        Just numericType ->
+          numericTypeIsIntegral numericType
+            && integerLiteralRangeFitsNumericType literalRange numericType
+        Nothing ->
+          False
+    (ConstraintTypeList signatureElementType, TListType elementType) ->
+      constraintSignatureTypeMatchesExpressionType state signatureElementType elementType
+    (ConstraintTypeTuple signatureElementTypes, TTupleType elementTypes)
+      | length signatureElementTypes == length elementTypes ->
+          and (zipWith (constraintSignatureTypeMatchesExpressionType state) signatureElementTypes elementTypes)
+    (ConstraintTypeApplication signatureName signatureArguments, TDataType typeName typeArguments)
+      | normalizeConstraintSignatureName (identifierText signatureName)
+          == normalizeConstraintSignatureName (identifierText typeName),
+        length signatureArguments == length typeArguments ->
+          and (zipWith (constraintSignatureTypeMatchesExpressionType state) signatureArguments typeArguments)
+    (ConstraintTypeFunction signatureArgument signatureResult, TFunctionType argumentType resultType) ->
+      constraintSignatureTypeMatchesExpressionType state signatureArgument argumentType
+        && constraintSignatureTypeMatchesExpressionType state signatureResult resultType
+    _ ->
+      case expressionTypeToRuntimeHint (defaultLiteralTypes (resolveType state expressionType)) of
+        Just argumentHint -> constraintSignatureTypesCompatible signatureType argumentHint
+        Nothing -> False
+
+numericTypeFromConstraintSignatureName :: Text -> Maybe NumericType
+numericTypeFromConstraintSignatureName =
+  numericTypeFromName . normalizeConstraintSignatureName
+
+concreteImplFactExists :: Text -> ConstraintSignatureType -> ScopeCapabilityFacts -> Bool
+concreteImplFactExists constraintName argumentHint facts =
+  any
+    (\candidateHint -> Set.member (constraintName <> "(" <> renderConstraintSignatureType candidateHint <> ")") (scopeConcreteImplFacts facts))
+    (constraintSignatureAliasVariants argumentHint)
+
+concreteImplMethodBodyExists :: Text -> ConstraintSignatureType -> ScopeCapabilityFacts -> Bool
+concreteImplMethodBodyExists methodKey argumentHint facts =
+  any
+    (\(ImplMethodType implTarget) -> constraintSignatureTypesCompatible implTarget argumentHint)
+    (Map.findWithDefault [] methodKey (scopeConcreteImplMethods facts))
+
+inferredEqualityConstraintCanUseStructuralRuntimeEquality :: InferState -> ScopeCapabilityFacts -> Maybe Text -> Text -> ExpressionType -> Bool
+inferredEqualityConstraintCanUseStructuralRuntimeEquality state facts maybeMethodKey constraintName argumentType =
+  maybeMethodKey == Nothing
+    && equalityConstraintNameCanUseStructuralRuntimeEquality state facts constraintName
+    && structuralRuntimeEqualityType state argumentType
+
+equalityConstraintNameCanUseStructuralRuntimeEquality :: InferState -> ScopeCapabilityFacts -> Text -> Bool
+equalityConstraintNameCanUseStructuralRuntimeEquality state facts constraintName =
+  activeEqualityClassName state == Just constraintName
+    || generatedHiddenEqualityClassFact constraintName facts
+
+generatedHiddenEqualityClassFact :: Text -> ScopeCapabilityFacts -> Bool
+generatedHiddenEqualityClassFact constraintName facts =
+  Set.member constraintName (scopeGeneratedEqualityClassFacts facts)
+    && Map.lookup constraintName (scopeClassFacts facts) == Just 1
+
+structuralRuntimeEqualityType :: InferState -> ExpressionType -> Bool
+structuralRuntimeEqualityType state argumentType =
+  case resolveType state argumentType of
+    TListType elementType ->
+      supportsRuntimeEqualityType state elementType
+    TTupleType elementTypes ->
+      all (supportsRuntimeEqualityType state) elementTypes
+    TDataType typeName typeArguments ->
+      supportsRuntimeEqualityType state (TDataType typeName typeArguments)
+    _ ->
+      False
+
+instantiateQualifiedMethodType :: Text -> InferState -> Maybe (Maybe ExpressionType, InferState)
+instantiateQualifiedMethodType nameText state =
+  case splitQualifiedMethodKey nameText of
+    Just (capabilityName, _)
+      | Map.member capabilityName (inferClassFacts state) ->
+          Just (resolveQualifiedMethodType nameText state)
+    _ -> Nothing
+
+resolveQualifiedMethodType :: Text -> InferState -> (Maybe ExpressionType, InferState)
+resolveQualifiedMethodType methodKey state =
+  case Map.lookup methodKey (inferClassMethodSignatures state) of
+    Nothing
+      | not (null (Map.findWithDefault [] methodKey (inferConcreteImplMethods state))) ->
+          (Nothing, state)
+      | otherwise ->
+          (Nothing, addTypeError state (mkMissingClassMethodError methodKey))
+    Just classMethodType ->
+      case Map.findWithDefault [] methodKey (inferConcreteImplMethods state) of
+        [] ->
+          (Nothing, addTypeError state (mkMissingImplMethodBodyError methodKey))
+        [implMethodType] ->
+          qualifiedMethodSignatureType methodKey classMethodType implMethodType state
+        _ ->
+          (Nothing, addTypeError state (mkAmbiguousQualifiedMethodBodyError methodKey))
+
+resolveQualifiedMethodApplicationType ::
+  Text ->
+  TypeEnv ->
+  InferState ->
+  [(Expr, ExpressionType)] ->
+  (Maybe ExpressionType, InferState)
+resolveQualifiedMethodApplicationType methodKey env state typedArguments =
+  case Map.lookup methodKey (inferClassMethodSignatures state) of
+    Nothing
+      | not (null (Map.findWithDefault [] methodKey (inferConcreteImplMethods state))) ->
+          (Nothing, state)
+      | otherwise ->
+          (Nothing, addTypeError state (mkMissingClassMethodError methodKey))
+    Just classMethodType ->
+      case inferQualifiedMethodRequirement methodKey classMethodType state argumentTypes of
+        Just inferredRequirement ->
+          inferredRequirement
+        Nothing ->
+          case Map.findWithDefault [] methodKey (inferConcreteImplMethods state) of
+            [] ->
+              (Nothing, addTypeError state (mkMissingImplMethodBodyError methodKey))
+            [implMethodType] ->
+              applyQualifiedMethodCandidateWithErrors methodKey classMethodType implMethodType state argumentTypes
+            implMethodTypes ->
+              selectQualifiedMethodCandidate methodKey classMethodType implMethodTypes env state typedArguments
+  where
+    argumentTypes = map snd typedArguments
+
+inferQualifiedMethodRequirement ::
+  Text ->
+  ClassMethodType ->
+  InferState ->
+  [ExpressionType] ->
+  Maybe (Maybe ExpressionType, InferState)
+inferQualifiedMethodRequirement methodKey (ClassMethodType classParameter methodSignature) state argumentTypes = do
+  (capabilityName, _) <- splitQualifiedMethodKey methodKey
+  classArity <- Map.lookup capabilityName (inferClassFacts state)
+  if classArity /= 1
+    then Nothing
+    else
+      if not (classMethodSignatureHasTargetArgument classParameter methodSignature)
+        then Nothing
+        else
+          let (classTarget, stateAfterClassTarget) = freshTypeVar state
+           in do
+                methodType <-
+                  classMethodPayloadToGenericExpressionType
+                    stateAfterClassTarget
+                    classParameter
+                    classTarget
+                    methodSignature
+                let (maybeResultType, stateAfterArguments) =
+                      applyKnownFunctionArguments methodType argumentTypes stateAfterClassTarget
+                    resolvedClassTarget = resolveType stateAfterArguments classTarget
+                case maybeResultType of
+                  Just resultType
+                    | not (Set.null (freeTypeVariables resolvedClassTarget)) ->
+                        Just
+                          ( Just resultType,
+                            addInferredMethodClassConstraint capabilityName methodKey resolvedClassTarget stateAfterArguments
+                          )
+                  _ ->
+                    Nothing
+
+classMethodSignatureHasTargetArgument :: Text -> SignaturePayload -> Bool
+classMethodSignatureHasTargetArgument classParameter methodSignature =
+  case signaturePayloadConstraintType methodSignature of
+    Just signatureType ->
+      let (argumentTypes, _) = constraintFunctionArgumentTypes signatureType
+       in any (constraintSignatureTypeContainsClassParameter classParameter) argumentTypes
+    Nothing ->
+      False
+
+selectQualifiedMethodCandidate ::
+  Text ->
+  ClassMethodType ->
+  [ImplMethodType] ->
+  TypeEnv ->
+  InferState ->
+  [(Expr, ExpressionType)] ->
+  (Maybe ExpressionType, InferState)
+selectQualifiedMethodCandidate methodKey classMethodType implMethodTypes env state typedArguments =
+  case preferredCandidates of
+    [] ->
+      ( Nothing,
+        addTypeError state (mkNoMatchingQualifiedMethodBodyError methodKey (resolvedArgumentTypes state))
+      )
+    [(matchedType, matchedState)] ->
+      (Just matchedType, matchedState)
+    _ ->
+      ( Nothing,
+        addTypeError state (mkAmbiguousQualifiedMethodBodyForArgumentsError methodKey (resolvedArgumentTypes state))
+      )
+  where
+    preferredCandidates =
+      case exactMatchingCandidates of
+        [] -> matchingCandidates
+        exactMatches -> exactMatches
+
+    exactMatchingCandidates =
+      filterExactMatches matchingCandidatesWithTargets
+
+    matchingCandidates =
+      map
+        (\(_, matchedType, matchedState) -> (matchedType, matchedState))
+        matchingCandidatesWithTargets
+
+    matchingCandidatesWithTargets =
+      foldr collectMatch [] implMethodTypes
+
+    collectMatch implMethodType matches =
+      case applyQualifiedMethodCandidate methodKey classMethodType implMethodType state argumentTypes of
+        (Just matchedType, matchedState) -> (implMethodType, matchedType, matchedState) : matches
+        (Nothing, _) -> matches
+
+    filterExactMatches candidates =
+      [ (matchedType, matchedState)
+        | (implMethodType, matchedType, matchedState) <- candidates,
+          qualifiedMethodCandidateExactlyMatchesArguments state env classMethodType implMethodType typedArguments
+      ]
+
+    resolvedArgumentTypes stateForRendering =
+      map (resolveType stateForRendering) argumentTypes
+
+    argumentTypes = map snd typedArguments
+
+qualifiedMethodCandidateExactlyMatchesArguments ::
+  InferState ->
+  TypeEnv ->
+  ClassMethodType ->
+  ImplMethodType ->
+  [(Expr, ExpressionType)] ->
+  Bool
+qualifiedMethodCandidateExactlyMatchesArguments state env (ClassMethodType classParameter methodSignature) (ImplMethodType implTarget) typedArguments =
+  case (signaturePayloadConstraintType methodSignature, substituteClassMethodSignature classParameter implTarget methodSignature) of
+    (Just genericSignature, Just substitutedSignature) ->
+      let (genericArgumentTypes, _) = constraintFunctionArgumentTypes genericSignature
+          (candidateArgumentTypes, _) = constraintFunctionArgumentTypes substitutedSignature
+          suppliedArgumentCount = length typedArguments
+          suppliedGenericArgumentTypes = take suppliedArgumentCount genericArgumentTypes
+          suppliedCandidateArgumentTypes = take suppliedArgumentCount candidateArgumentTypes
+          targetArgumentPositions =
+            map (constraintSignatureTypeContainsClassParameter classParameter) suppliedGenericArgumentTypes
+       in suppliedArgumentCount <= length genericArgumentTypes
+            && suppliedArgumentCount <= length candidateArgumentTypes
+            && or targetArgumentPositions
+            && and
+              ( zipWith3
+                  exactCandidateArgumentMatches
+                  targetArgumentPositions
+                  suppliedCandidateArgumentTypes
+                  typedArguments
+              )
+    _ ->
+      False
+  where
+    exactCandidateArgumentMatches targetArgumentPosition signatureType (argumentExpr, expressionType) =
+      not targetArgumentPosition
+        || constraintSignatureTypeExactlyMatchesExpressionType state signatureType expressionType
+          && constraintSignatureExpressionHasExactEvidence env signatureType argumentExpr
+
+constraintSignatureExpressionHasExactEvidence :: TypeEnv -> ConstraintSignatureType -> Expr -> Bool
+constraintSignatureExpressionHasExactEvidence env signatureType argumentExpr =
+  case (signatureType, argumentExpr) of
+    (ConstraintTypeList elementType, EList elements) ->
+      not (null elements)
+        && all (constraintSignatureExpressionHasExactEvidence env elementType) elements
+    (ConstraintTypeTuple elementTypes, ETuple elements)
+      | length elementTypes == length elements ->
+          and (zipWith (constraintSignatureExpressionHasExactEvidence env) elementTypes elements)
+    (ConstraintTypeApplication typeName typeArguments, EApply {}) ->
+      constructorApplicationExpressionHasExactEvidence env typeName typeArguments argumentExpr
+        || constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (ConstraintTypeFunction {}, _) ->
+      constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (_, EVar {})
+      | constraintSignatureTypeContainsList signatureType ->
+          constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (_, EApply {})
+      | constraintSignatureTypeContainsList signatureType ->
+          constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (_, EIf {}) ->
+      constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (_, EPatternCase {}) ->
+      constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    (_, EBlock {})
+      | constraintSignatureTypeContainsList signatureType ->
+          constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr
+    _ -> True
+
+constraintSignatureExpressionRuntimeHintMatches :: TypeEnv -> ConstraintSignatureType -> Expr -> Bool
+constraintSignatureExpressionRuntimeHintMatches env signatureType argumentExpr =
+  case constraintSignatureExpressionRuntimeHint env argumentExpr of
+    Just runtimeHint -> runtimeHint == signatureType
+    Nothing -> False
+
+constraintSignatureExpressionRuntimeHint :: TypeEnv -> Expr -> Maybe ConstraintSignatureType
+constraintSignatureExpressionRuntimeHint env argumentExpr =
+  constraintSignatureExpressionRuntimeHintWithLocalHints env Map.empty argumentExpr
+
+constraintSignatureExpressionRuntimeHintWithLocalHints ::
+  TypeEnv ->
+  Map Text ConstraintSignatureType ->
+  Expr ->
+  Maybe ConstraintSignatureType
+constraintSignatureExpressionRuntimeHintWithLocalHints env localHints argumentExpr =
+  case argumentExpr of
+    EVar referencedName ->
+      Map.lookup (identifierText referencedName) localHints
+        <|> (Map.lookup referencedName env >>= typeBindingRuntimeHint)
+    EApply (EApply dollarExpr functionExpr) _
+      | builtinDollarOperatorExpr env dollarExpr ->
+          case constraintSignatureExpressionRuntimeHintWithLocalHints env localHints functionExpr of
+            Just (ConstraintTypeFunction _ resultType) -> Just resultType
+            _ -> Nothing
+    EApply functionExpr _ ->
+      case constraintSignatureExpressionRuntimeHintWithLocalHints env localHints functionExpr of
+        Just (ConstraintTypeFunction _ resultType) -> Just resultType
+        _ -> Nothing
+    EIf _ thenExpr elseExpr ->
+      commonConstraintSignatureExpressionRuntimeHint env localHints [thenExpr, elseExpr]
+    EPatternCase _ caseArms ->
+      commonConstraintSignatureExpressionRuntimeHint env localHints [bodyExpr | CaseArm _ _ bodyExpr <- caseArms]
+    EBlock statements ->
+      constraintSignatureBlockRuntimeHint env localHints statements
+    _ -> Nothing
+
+commonConstraintSignatureExpressionRuntimeHint ::
+  TypeEnv ->
+  Map Text ConstraintSignatureType ->
+  [Expr] ->
+  Maybe ConstraintSignatureType
+commonConstraintSignatureExpressionRuntimeHint _ _ [] = Nothing
+commonConstraintSignatureExpressionRuntimeHint env localHints (firstExpr : restExprs) = do
+  firstHint <- constraintSignatureExpressionRuntimeHintWithLocalHints env localHints firstExpr
+  if all
+    (\expr -> constraintSignatureExpressionRuntimeHintWithLocalHints env localHints expr == Just firstHint)
+    restExprs
+    then Just firstHint
+    else Nothing
+
+constraintSignatureBlockRuntimeHint ::
+  TypeEnv ->
+  Map Text ConstraintSignatureType ->
+  [Statement] ->
+  Maybe ConstraintSignatureType
+constraintSignatureBlockRuntimeHint env initialLocalHints statements =
+  go initialLocalHints Map.empty statements
+  where
+    go localHints _ [] =
+      Nothing
+    go localHints _ [SExpr _ expr] =
+      constraintSignatureExpressionRuntimeHintWithLocalHints env localHints expr
+    go localHints pendingHints (statement : rest) =
+      case statement of
+        SSignature name _ signaturePayload ->
+          let nameText = identifierText name
+              nextPendingHints =
+                case signaturePayloadRuntimeHint signaturePayload of
+                  Just runtimeHint -> Map.insert nameText runtimeHint pendingHints
+                  Nothing -> Map.delete nameText pendingHints
+           in go localHints nextPendingHints rest
+        SLet name _ valueExpr ->
+          let nameText = identifierText name
+              bindingHint =
+                Map.lookup nameText pendingHints
+                  <|> constraintSignatureExpressionRuntimeHintWithLocalHints env localHints valueExpr
+              nextLocalHints =
+                case bindingHint of
+                  Just runtimeHint -> Map.insert nameText runtimeHint localHints
+                  Nothing -> localHints
+           in go nextLocalHints (Map.delete nameText pendingHints) rest
+        _ ->
+          go localHints pendingHints rest
+
+signaturePayloadRuntimeHint :: SignaturePayload -> Maybe ConstraintSignatureType
+signaturePayloadRuntimeHint signaturePayload =
+  case signaturePayload of
+    SignatureType signatureType ->
+      expressionTypeToRuntimeHint (signatureTypeToExpressionType signatureType)
+    ConstrainedSignature _ signatureType
+      | Set.null (constraintSignatureTypeVariableNames signatureType) ->
+          Just signatureType
+    ConstrainedSignature _ signatureType ->
+      constraintSignatureTypeToExpressionType signatureType >>= expressionTypeToRuntimeHint
+    UnsupportedSignature {} ->
+      Nothing
+
+typeBindingRuntimeHint :: TypeBinding -> Maybe ConstraintSignatureType
+typeBindingRuntimeHint binding =
+  case binding of
+    PlainTypeBinding bindingType ->
+      expressionTypeToRuntimeHint (defaultLiteralTypes bindingType)
+    SchemeTypeBinding typeScheme
+      | Set.null (schemeQuantifiedVariables typeScheme) ->
+          expressionTypeToRuntimeHint (defaultLiteralTypes (schemeResultType typeScheme))
+    OperatorAliasSchemeTypeBinding _ typeScheme
+      | Set.null (schemeQuantifiedVariables typeScheme) ->
+          expressionTypeToRuntimeHint (defaultLiteralTypes (schemeResultType typeScheme))
+    _ -> Nothing
+
+constraintSignatureTypeContainsList :: ConstraintSignatureType -> Bool
+constraintSignatureTypeContainsList signatureType =
+  case signatureType of
+    ConstraintTypeList {} -> True
+    ConstraintTypeTuple elementTypes ->
+      any constraintSignatureTypeContainsList elementTypes
+    ConstraintTypeApplication _ typeArguments ->
+      any constraintSignatureTypeContainsList typeArguments
+    ConstraintTypeFunction argumentType resultType ->
+      constraintSignatureTypeContainsList argumentType
+        || constraintSignatureTypeContainsList resultType
+    ConstraintTypeName {} -> False
+
+constructorApplicationExpressionHasExactEvidence :: TypeEnv -> Name -> [ConstraintSignatureType] -> Expr -> Bool
+constructorApplicationExpressionHasExactEvidence env typeName typeArguments argumentExpr =
+  case constructorExpressionSpine argumentExpr of
+    Just (constructorName, constructorArgumentExprs) ->
+      case Map.lookup constructorName env of
+        Just (ConstructorTypeBinding constructorTypeName typeParameters constructorArgumentTypes)
+          | constructorTypeName == typeName,
+            length typeParameters == length typeArguments,
+            length constructorArgumentTypes == length constructorArgumentExprs ->
+              let typeParameterBindings =
+                    Map.fromList (zip (map identifierText typeParameters) typeArguments)
+               in and
+                    ( zipWith
+                        (constructorArgumentExpressionHasExactEvidence env typeParameterBindings)
+                        constructorArgumentTypes
+                        constructorArgumentExprs
+                    )
+        _ -> False
+    Nothing -> False
+
+constructorExpressionSpine :: Expr -> Maybe (Name, [Expr])
+constructorExpressionSpine expr =
+  go [] expr
+  where
+    go argumentExprs currentExpr =
+      case currentExpr of
+        EApply functionExpr argumentExpr ->
+          go (argumentExpr : argumentExprs) functionExpr
+        EVar constructorName ->
+          Just (constructorName, argumentExprs)
+        _ ->
+          Nothing
+
+constructorArgumentExpressionHasExactEvidence :: TypeEnv -> Map Text ConstraintSignatureType -> ConstructorArgumentType -> Expr -> Bool
+constructorArgumentExpressionHasExactEvidence env typeParameterBindings constructorArgument argumentExpr =
+  case constructorArgument of
+    ConstructorArgumentParameter parameterName ->
+      case Map.lookup parameterName typeParameterBindings of
+        Just concreteArgumentType ->
+          constraintSignatureExpressionHasExactEvidence env concreteArgumentType argumentExpr
+        Nothing ->
+          True
+    ConstructorArgumentMonomorphic {} ->
+      True
+    ConstructorArgumentFresh ->
+      True
+
+constraintSignatureTypeExactlyMatchesExpressionType :: InferState -> ConstraintSignatureType -> ExpressionType -> Bool
+constraintSignatureTypeExactlyMatchesExpressionType state signatureType expressionType =
+  case constraintSignatureTypeToExpressionTypeWithState state Map.empty signatureType of
+    Just signatureExpressionType ->
+      resolveType state signatureExpressionType == defaultLiteralTypes (resolveType state expressionType)
+    Nothing ->
+      False
+
+applyQualifiedMethodCandidate ::
+  Text ->
+  ClassMethodType ->
+  ImplMethodType ->
+  InferState ->
+  [ExpressionType] ->
+  (Maybe ExpressionType, InferState)
+applyQualifiedMethodCandidate methodKey classMethodType implMethodType state argumentTypes =
+  case qualifiedMethodSignatureType methodKey classMethodType implMethodType state of
+    (Nothing, nextState) ->
+      (Nothing, nextState)
+    (Just methodType, stateAfterMethodType) ->
+      applyKnownFunctionArguments methodType argumentTypes stateAfterMethodType
+
+applyQualifiedMethodCandidateWithErrors ::
+  Text ->
+  ClassMethodType ->
+  ImplMethodType ->
+  InferState ->
+  [ExpressionType] ->
+  (Maybe ExpressionType, InferState)
+applyQualifiedMethodCandidateWithErrors methodKey classMethodType implMethodType state argumentTypes =
+  case qualifiedMethodSignatureType methodKey classMethodType implMethodType state of
+    (Nothing, nextState) ->
+      (Nothing, nextState)
+    (Just methodType, stateAfterMethodType) ->
+      applyKnownFunctionArgumentsWithErrors methodType argumentTypes stateAfterMethodType
+
+applyKnownFunctionArguments ::
+  ExpressionType ->
+  [ExpressionType] ->
+  InferState ->
+  (Maybe ExpressionType, InferState)
+applyKnownFunctionArguments functionType argumentTypes state =
+  foldl' step (Just functionType, state) argumentTypes
+  where
+    step (Nothing, stateAcc) _ =
+      (Nothing, stateAcc)
+    step (Just currentFunctionType, stateAcc) argumentType =
+      let (resultTypeVar, stateWithResultVar) = freshTypeVar stateAcc
+       in case unifyTypes currentFunctionType (TFunctionType argumentType resultTypeVar) stateWithResultVar of
+            Just unifiedState ->
+              (Just (resolveType unifiedState resultTypeVar), unifiedState)
+            Nothing ->
+              (Nothing, stateAcc)
+
+applyKnownFunctionArgumentsWithErrors ::
+  ExpressionType ->
+  [ExpressionType] ->
+  InferState ->
+  (Maybe ExpressionType, InferState)
+applyKnownFunctionArgumentsWithErrors functionType argumentTypes state =
+  foldl' step (Just functionType, state) argumentTypes
+  where
+    step (Nothing, stateAcc) _ =
+      (Nothing, stateAcc)
+    step (Just currentFunctionType, stateAcc) argumentType =
+      let (resultTypeVar, stateWithResultVar) = freshTypeVar stateAcc
+       in case unifyTypes currentFunctionType (TFunctionType argumentType resultTypeVar) stateWithResultVar of
+            Just unifiedState ->
+              (Just (resolveType unifiedState resultTypeVar), unifiedState)
+            Nothing ->
+              ( Nothing,
+                addTypeError
+                  stateWithResultVar
+                  ( mkApplyTypeError
+                      (resolveType stateWithResultVar currentFunctionType)
+                      (resolveType stateWithResultVar argumentType)
+                  )
+              )
+
+qualifiedMethodSignatureType ::
+  Text ->
+  ClassMethodType ->
+  ImplMethodType ->
+  InferState ->
+  (Maybe ExpressionType, InferState)
+qualifiedMethodSignatureType methodKey (ClassMethodType classParameter methodSignature) (ImplMethodType implTarget) state =
+  case classMethodPayloadToExpressionType state classParameter implTarget methodSignature of
+    Just methodType -> (Just methodType, state)
+    Nothing ->
+      (Nothing, addTypeError state (mkInvalidQualifiedMethodSignatureError methodKey methodSignature))
+
+classMethodPayloadToExpressionType ::
+  InferState ->
+  Text ->
+  ConstraintSignatureType ->
+  SignaturePayload ->
+  Maybe ExpressionType
+classMethodPayloadToExpressionType state classParameter implTarget methodSignature =
+  substituteClassMethodSignature classParameter implTarget methodSignature
+    >>= constraintSignatureTypeToExpressionTypeWithState state Map.empty
+
+classMethodPayloadToGenericExpressionType ::
+  InferState ->
+  Text ->
+  ExpressionType ->
+  SignaturePayload ->
+  Maybe ExpressionType
+classMethodPayloadToGenericExpressionType state classParameter classTarget methodSignature =
+  signaturePayloadConstraintType methodSignature
+    >>= constraintSignatureTypeToExpressionTypeWithState
+      state
+      (Map.singleton classParameter classTarget)
+
+constraintSignatureTypeToExpressionTypeWithState ::
+  InferState ->
+  Map Text ExpressionType ->
+  ConstraintSignatureType ->
+  Maybe ExpressionType
+constraintSignatureTypeToExpressionTypeWithState state signatureVariables signatureType =
+  case signatureType of
+    ConstraintTypeName name ->
+      case identifierText name of
+        "Int" -> Just TIntType
+        "Float" -> Just TFloatType
+        "Bool" -> Just TBoolType
+        typeName ->
+          case numericTypeNameToExpressionType typeName of
+            Just numericType -> Just numericType
+            Nothing ->
+              case Map.lookup typeName signatureVariables of
+                Just variableType -> Just variableType
+                Nothing
+                  | Map.member typeName (inferDataTypes state) ->
+                      Just (TDataType name [])
+                  | otherwise ->
+                      Nothing
+    ConstraintTypeApplication name arguments
+      | Map.member (identifierText name) (inferDataTypes state) ->
+          TDataType name <$> traverse (constraintSignatureTypeToExpressionTypeWithState state signatureVariables) arguments
+      | otherwise ->
+          Nothing
+    ConstraintTypeList innerType ->
+      TListType <$> constraintSignatureTypeToExpressionTypeWithState state signatureVariables innerType
+    ConstraintTypeTuple elementTypes ->
+      TTupleType <$> traverse (constraintSignatureTypeToExpressionTypeWithState state signatureVariables) elementTypes
+    ConstraintTypeFunction argumentType resultType ->
+      TFunctionType
+        <$> constraintSignatureTypeToExpressionTypeWithState state signatureVariables argumentType
+        <*> constraintSignatureTypeToExpressionTypeWithState state signatureVariables resultType
+
+freshTypeVars :: Int -> InferState -> ([ExpressionType], InferState)
+freshTypeVars count initialState =
+  go count [] initialState
+  where
+    go remaining acc state
+      | remaining <= 0 = (reverse acc, state)
+      | otherwise =
+          let (typeVar, nextState) = freshTypeVar state
+           in go (remaining - 1) (typeVar : acc) nextState
+
+data SignaturePayloadType = SignaturePayloadType
+  { signaturePayloadDeclaredType :: ExpressionType,
+    signaturePayloadExplicitConstraints :: [TypeSchemeConstraint],
+    signaturePayloadVariableOrder :: [Int]
+  }
+
+-- | Normalize the currently accepted signature subset. Unsupported surfaces
+-- return `Nothing` so callers can emit the stable signature diagnostic.
+signaturePayloadToSignatureType :: SignaturePayload -> InferState -> (Maybe SignaturePayloadType, InferState)
+signaturePayloadToSignatureType signaturePayload state =
+  case signaturePayload of
+    SignatureType signatureType ->
+      (Just (SignaturePayloadType (signatureTypeToExpressionType signatureType) [] []), state)
+    ConstrainedSignature [] signatureType ->
+      (fmap (\declaredType -> SignaturePayloadType declaredType [] []) (constraintSignatureTypeToExpressionType signatureType), state)
+    ConstrainedSignature constraints signatureType
+      | supportedConcreteConstraints state constraints ->
+          (fmap (\declaredType -> SignaturePayloadType declaredType [] []) (constraintSignatureTypeToExpressionType signatureType), state)
+      | supportedVariableConstraints state constraints signatureType ->
+          variableConstraintSignaturePayloadToExpressionType constraints signatureType state
+      | otherwise ->
+          (Nothing, state)
+    UnsupportedSignature {} ->
+      (Nothing, state)
+
+signatureTypeToExpressionType :: SignatureType -> ExpressionType
+signatureTypeToExpressionType signatureType =
+  case signatureType of
+    TypeInt -> TIntType
+    TypeFloat -> TFloatType
+    TypeNumeric numericType -> TNumericType numericType
+    TypeBool -> TBoolType
+    TypeList innerType ->
+      TListType (signatureTypeToExpressionType innerType)
+    TypeTuple elementTypes ->
+      TTupleType (map signatureTypeToExpressionType elementTypes)
+    TypeFunction argumentType resultType ->
+      TFunctionType
+        (signatureTypeToExpressionType argumentType)
+        (signatureTypeToExpressionType resultType)
+
+constraintSignatureTypeToExpressionType :: ConstraintSignatureType -> Maybe ExpressionType
+constraintSignatureTypeToExpressionType =
+  constraintSignatureTypeToExpressionTypeWithVariables Map.empty
+
+constraintSignatureTypeToExpressionTypeWithVariables ::
+  Map Text ExpressionType ->
+  ConstraintSignatureType ->
+  Maybe ExpressionType
+constraintSignatureTypeToExpressionTypeWithVariables signatureVariables signatureType =
+  case signatureType of
+    ConstraintTypeName name ->
+      case identifierText name of
+        "Int" -> Just TIntType
+        "Float" -> Just TFloatType
+        "Bool" -> Just TBoolType
+        typeName ->
+          case numericTypeNameToExpressionType typeName of
+            Just numericType -> Just numericType
+            Nothing -> Map.lookup typeName signatureVariables
+    ConstraintTypeApplication {} ->
+      Nothing
+    ConstraintTypeList innerType ->
+      TListType <$> constraintSignatureTypeToExpressionTypeWithVariables signatureVariables innerType
+    ConstraintTypeTuple elementTypes ->
+      TTupleType <$> traverse (constraintSignatureTypeToExpressionTypeWithVariables signatureVariables) elementTypes
+    ConstraintTypeFunction argumentType resultType ->
+      TFunctionType
+        <$> constraintSignatureTypeToExpressionTypeWithVariables signatureVariables argumentType
+        <*> constraintSignatureTypeToExpressionTypeWithVariables signatureVariables resultType
+
+variableConstraintSignaturePayloadToExpressionType ::
+  [SignatureConstraint] ->
+  ConstraintSignatureType ->
+  InferState ->
+  (Maybe SignaturePayloadType, InferState)
+variableConstraintSignaturePayloadToExpressionType constraints signatureType state =
+  let variableNames = constraintSignatureTypeVariableNamesInOrder signatureType
+      (signatureVariables, nextState) = allocateSignatureTypeVariables variableNames state
+      convertedType =
+        constraintSignatureTypeToExpressionTypeWithVariables signatureVariables signatureType
+      convertedConstraints =
+        traverse (variableConstraintToTypeSchemeConstraint signatureVariables) constraints
+      variableOrder =
+        [ typeVar
+          | variableName <- variableNames,
+            Just (TVarType typeVar) <- [Map.lookup variableName signatureVariables]
+        ]
+   in
+    case (convertedType, convertedConstraints) of
+      (Just expressionType, Just explicitConstraints) ->
+        (Just (SignaturePayloadType expressionType explicitConstraints variableOrder), nextState)
+      _ -> (Nothing, state)
+
+variableConstraintToTypeSchemeConstraint ::
+  Map Text ExpressionType ->
+  SignatureConstraint ->
+  Maybe TypeSchemeConstraint
+variableConstraintToTypeSchemeConstraint signatureVariables (SignatureConstraint constraintName arguments) =
+  case arguments of
+    [ConstraintTypeName argumentName] ->
+      TypeSchemeConstraint (identifierText constraintName)
+        <$> Map.lookup (identifierText argumentName) signatureVariables
+    _ -> Nothing
+
+allocateSignatureTypeVariables :: [Text] -> InferState -> (Map Text ExpressionType, InferState)
+allocateSignatureTypeVariables variableNames state =
+  foldl' allocate (Map.empty, state) variableNames
+  where
+    allocate (signatureVariables, stateAcc) variableName =
+      let (variableType, nextState) = freshTypeVar stateAcc
+       in (Map.insert variableName variableType signatureVariables, nextState)
+
+supportedConcreteConstraints :: InferState -> [SignatureConstraint] -> Bool
+supportedConcreteConstraints state constraints =
+  not (null constraints)
+    && isNothing (duplicateConstraintName constraints)
+    && all (supportedConcreteConstraint state) constraints
+
+-- | Variable constrained signatures are accepted when every constrained
+-- variable appears in the body; extra body variables remain unconstrained.
+supportedVariableConstraints :: InferState -> [SignatureConstraint] -> ConstraintSignatureType -> Bool
+supportedVariableConstraints state constraints signatureType =
+  not (null constraints)
+    && isNothing (duplicateConstraintName constraints)
+    && all (supportedVariableConstraint state) constraints
+    && constraintSignatureTypeSupportsVariableBody signatureType
+    && not (Set.null signatureVariableNames)
+    && constraintVariableNames `Set.isSubsetOf` signatureVariableNames
+  where
+    signatureVariableNames =
+      constraintSignatureTypeVariableNames signatureType
+    constraintVariableNames =
+      Set.unions (map constraintVariableNamesInSupportedConstraint constraints)
+
+supportedConcreteConstraint :: InferState -> SignatureConstraint -> Bool
+supportedConcreteConstraint state (SignatureConstraint constraintName arguments) =
+  case (Map.lookup (identifierText constraintName) (inferClassFacts state), arguments) of
+    (Just 1, [argument]) ->
+      concreteConstraintArgument argument
+        && Set.member
+          (constraintImplFactKey constraintName argument)
+          (inferConcreteImplFacts state)
+    _ -> False
+
+supportedVariableConstraint :: InferState -> SignatureConstraint -> Bool
+supportedVariableConstraint state (SignatureConstraint constraintName arguments) =
+  case (Map.lookup (identifierText constraintName) (inferClassFacts state), arguments) of
+    (Just 1, [ConstraintTypeName argumentName]) ->
+      identifierLooksLikeTypeVariable argumentName
+    _ -> False
+
+constraintVariableNamesInSupportedConstraint :: SignatureConstraint -> Set Text
+constraintVariableNamesInSupportedConstraint constraint =
+  case constraint of
+    SignatureConstraint _ [ConstraintTypeName argumentName]
+      | identifierLooksLikeTypeVariable argumentName ->
+          Set.singleton (identifierText argumentName)
+    _ -> Set.empty
+
+constraintSignatureTypeVariableNames :: ConstraintSignatureType -> Set Text
+constraintSignatureTypeVariableNames signatureType =
+  case signatureType of
+    ConstraintTypeName name
+      | identifierLooksLikeTypeVariable name ->
+          Set.singleton (identifierText name)
+      | otherwise ->
+          Set.empty
+    ConstraintTypeApplication _ arguments ->
+      Set.unions (map constraintSignatureTypeVariableNames arguments)
+    ConstraintTypeList innerType ->
+      constraintSignatureTypeVariableNames innerType
+    ConstraintTypeTuple elementTypes ->
+      Set.unions (map constraintSignatureTypeVariableNames elementTypes)
+    ConstraintTypeFunction argumentType resultType ->
+      Set.union
+        (constraintSignatureTypeVariableNames argumentType)
+        (constraintSignatureTypeVariableNames resultType)
+
+constraintSignatureTypeSupportsVariableBody :: ConstraintSignatureType -> Bool
+constraintSignatureTypeSupportsVariableBody signatureType =
+  case signatureType of
+    ConstraintTypeName {} -> True
+    ConstraintTypeApplication {} -> False
+    ConstraintTypeList innerType ->
+      constraintSignatureTypeSupportsVariableBody innerType
+    ConstraintTypeTuple elementTypes ->
+      all constraintSignatureTypeSupportsVariableBody elementTypes
+    ConstraintTypeFunction argumentType resultType ->
+      constraintSignatureTypeSupportsVariableBody argumentType
+        && constraintSignatureTypeSupportsVariableBody resultType
+
+numericTypeNameToExpressionType :: Text -> Maybe ExpressionType
+numericTypeNameToExpressionType typeName =
+  TNumericType <$> numericTypeFromName typeName
+
+defaultLiteralTypes :: ExpressionType -> ExpressionType
+defaultLiteralTypes =
+  defaultLiteralTypesWith TIntType
+
+defaultBindingLiteralTypes :: ExpressionType -> ExpressionType
+defaultBindingLiteralTypes =
+  defaultLiteralTypesWith (TNumericType NumericInt64)
+
+defaultLiteralTypesWith :: ExpressionType -> ExpressionType -> ExpressionType
+defaultLiteralTypesWith integerLiteralDefault expressionType =
+  case expressionType of
+    TIntegerLiteralType {} -> integerLiteralDefault
+    TListType elementType ->
+      TListType (defaultLiteralTypesWith integerLiteralDefault elementType)
+    TTupleType elementTypes ->
+      TTupleType (map (defaultLiteralTypesWith integerLiteralDefault) elementTypes)
+    TDataType typeName typeArguments ->
+      TDataType typeName (map (defaultLiteralTypesWith integerLiteralDefault) typeArguments)
+    TFunctionType inputType outputType ->
+      TFunctionType
+        (defaultLiteralTypesWith integerLiteralDefault inputType)
+        (defaultLiteralTypesWith integerLiteralDefault outputType)
+    _ -> expressionType
+
+expressionTypeToRuntimeHint :: ExpressionType -> Maybe ConstraintSignatureType
+expressionTypeToRuntimeHint expressionType =
+  case expressionType of
+    TIntType -> Just (ConstraintTypeName "Int")
+    TIntegerLiteralType literalRange
+      | integerLiteralRangeFitsNumericType literalRange NumericInt64 ->
+          Just (ConstraintTypeName "Int")
+      | otherwise -> Nothing
+    TFloatType -> Just (ConstraintTypeName "Float")
+    TNumericType numericType -> Just (ConstraintTypeName (sourceName (mkIdentifier (renderNumericTypeName numericType))))
+    TBoolType -> Just (ConstraintTypeName "Bool")
+    TListType elementType ->
+      ConstraintTypeList <$> expressionTypeToRuntimeHint elementType
+    TTupleType elementTypes ->
+      ConstraintTypeTuple <$> traverse expressionTypeToRuntimeHint elementTypes
+    TDataType typeName typeArguments ->
+      case traverse expressionTypeToRuntimeHint typeArguments of
+        Just [] -> Just (ConstraintTypeName typeName)
+        Just argumentHints ->
+          Just (ConstraintTypeApplication typeName argumentHints)
+        Nothing -> Nothing
+    TFunctionType inputType outputType ->
+      ConstraintTypeFunction
+        <$> expressionTypeToRuntimeHint inputType
+        <*> expressionTypeToRuntimeHint outputType
+    TVarType {} -> Nothing
+
+expressionTypeToRuntimeHintWithVariables :: Map Int ConstraintSignatureType -> ExpressionType -> Maybe ConstraintSignatureType
+expressionTypeToRuntimeHintWithVariables variableHints expressionType =
+  case expressionType of
+    TIntType -> Just (ConstraintTypeName "Int")
+    TIntegerLiteralType literalRange
+      | integerLiteralRangeFitsNumericType literalRange NumericInt64 ->
+          Just (ConstraintTypeName "Int")
+      | otherwise -> Nothing
+    TFloatType -> Just (ConstraintTypeName "Float")
+    TNumericType numericType -> Just (ConstraintTypeName (sourceName (mkIdentifier (renderNumericTypeName numericType))))
+    TBoolType -> Just (ConstraintTypeName "Bool")
+    TListType elementType ->
+      ConstraintTypeList <$> expressionTypeToRuntimeHintWithVariables variableHints elementType
+    TTupleType elementTypes ->
+      ConstraintTypeTuple <$> traverse (expressionTypeToRuntimeHintWithVariables variableHints) elementTypes
+    TDataType typeName typeArguments ->
+      case traverse (expressionTypeToRuntimeHintWithVariables variableHints) typeArguments of
+        Just [] -> Just (ConstraintTypeName typeName)
+        Just argumentHints ->
+          Just (ConstraintTypeApplication typeName argumentHints)
+        Nothing -> Nothing
+    TFunctionType inputType outputType ->
+      ConstraintTypeFunction
+        <$> expressionTypeToRuntimeHintWithVariables variableHints inputType
+        <*> expressionTypeToRuntimeHintWithVariables variableHints outputType
+    TVarType typeVar ->
+      Map.lookup typeVar variableHints
+
+addInferredClassConstraint :: Text -> ExpressionType -> InferState -> InferState
+addInferredClassConstraint constraintName argumentType state =
+  modifyInferenceOutput
+    ( \output ->
+        output
+          { outputInferredConstraints =
+              TypeSchemeInferredConstraint constraintName argumentType : inferInferredClassConstraints state
+          }
+    )
+    state
+
+addInferredMethodClassConstraint :: Text -> Text -> ExpressionType -> InferState -> InferState
+addInferredMethodClassConstraint constraintName methodKey argumentType state =
+  modifyInferenceOutput
+    ( \output ->
+        output
+          { outputInferredConstraints =
+              TypeSchemeMethodConstraint constraintName methodKey argumentType : inferInferredClassConstraints state
+          }
+    )
+    state
+
+addInferredEqualityClassConstraintIfVisible :: ExpressionType -> InferState -> InferState
+addInferredEqualityClassConstraintIfVisible argumentType state =
+  case activeEqualityClassName state of
+    Just equalityClassName -> addInferredClassConstraint equalityClassName argumentType state
+    Nothing -> state
+
+activeEqualityClassName :: InferState -> Maybe Text
+activeEqualityClassName state =
+  case inferCurrentModulePath state of
+    Just modulePath
+      | let replayQualifiedEqName = moduleReplayQualifiedName modulePath "Eq",
+        classFactIsUnary replayQualifiedEqName ->
+          Just replayQualifiedEqName
+    _ ->
+      if classFactIsUnary "Eq"
+        then Just "Eq"
+        else Nothing
+  where
+    classFactIsUnary className =
+      Map.lookup className (inferClassFacts state) == Just 1
+
+moduleReplayQualifiedName :: [Text] -> Text -> Text
+moduleReplayQualifiedName modulePath name =
+  qualifiedIdentifierText "__module" (Text.intercalate "::" modulePath <> "::" <> name)
